# docs/openapi/openapi-2.0.0.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionCreate": {
        "properties": {
          "domain": {
            "default": "political",
            "title": "Domain",
            "type": "string"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "resolution_rule": {
            "additionalProperties": true,
            "default": {
              "type": "manual"
            },
            "title": "Resolution Rule",
            "type": "object"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "key",
          "text"
        ],
        "title": "QuestionCreate",
        "type": "object"
      },
      "ResolveBody": {
        "properties": {
          "outcome": {
            "title": "Outcome",
            "type": "integer"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "ResolveBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "OIDC/JWT bearer token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Conversational copilot over the cassandra-core engines. All numbers come from scored engines; the LLM (optional) only parses intents.",
    "title": "ARGUS Strategic Intelligence Copilot",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/questions": {
      "get": {
        "operationId": "list_questions_v1_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "resolved",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resolved"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "question-registry"
        ]
      },
      "post": {
        "operationId": "create_question_v1_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Question",
        "tags": [
          "question-registry"
        ]
      }
    },
...
    "/v1/questions/{key}": {
      "get": {
        "operationId": "get_question_v1_questions__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question",
        "tags": [
          "question-registry"
        ]
      }
    },
    "/v1/questions/{key}/resolve": {
      "post": {
        "operationId": "resolve_question_v1_questions__key__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Question",
        "tags": [
          "question-registry"
        ]
      }
    },
...
  }
}